+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Release Notes
 Forthcoming
 -----------
 * [demos] display modes demonstrating usage of various options, `#275 <https://github.com/splintered-reality/py_trees/pull/275>`_
+* [display] enforce left to right ordering of children in dot graphs, `#276 <https://github.com/splintered-reality/py_trees/pull/276>`_
 
 2.0.8 (2020-02-01)
 ------------------